Move socket to instance variable
Makes it work on Firefox. Not entirely sure why, but the socket does not
need to be in React state and I would say it's best practice for it not
to be, since it does not affect what's displayed and changes/updates to
it does not need to trigger a rerender. Some citations:
https://reactjs.org/docs/react-component.html#state
https://reactjs.org/docs/state-and-lifecycle.html#adding-lifecycle-methods-to-a-class
https://www.freecodecamp.org/news/where-do-i-belong-a-guide-to-saving-react-component-data-in-state-store-static-and-this-c49b335e2a00/